Insert default options on datePicker
This is more a personal preference, you may not like default options, but if you think about this concrete use case, it makes sense. Always when we work with dates, we normally end using `new Date()` by default.

I think it is a good idea to provide that default here because I think that the most common use is that one.

What do you think? That would need an update on the docs but since it is another branch, I don't think I can do that at once.