BatteryStatus example not working. Fixed by adding correct parameter
This example was not working.
Listener $cordovaBatteryStatus.$on('batterystatus', function (result) is receiving incorrect number of parameter to make example work. result is only the event that was fired. info is the correct parameter that contains the level and isPlugged attributes. 
Fixed example by adding the info parameter: $cordovaBatteryStatus.$on('batterystatus', function (result, info) {
Tested on a real device (iPhone and Android)